+++ test/SP1VerifyRv.t.sol
@@ -21,8 +21,10 @@ struct SP1ProofFixtureJson {
 
 contract UpdatedRvVerifiverTest is Test {
     using stdJson for string;
+
     uint64 constant fraction_bits = 40;
     RvVerifier public rvVerifier;
+
     using SignedMath for int64;
 
     function loadFixture() public view returns (SP1ProofFixtureJson memory) {
@@ -52,40 +54,38 @@ contract UpdatedRvVerifiverTest is Test {
 
     function test_ValidUpdatedRvProof() public view {
         SP1ProofFixtureJson memory fixture = loadFixture();
-        (bytes8 n_inv_sqrt, bytes8 n1_inv, bytes8 s2, bytes8 n_bytes, bytes32 digest) = rvVerifier.verifyRvProof(
-            fixture.proof,
-            fixture.publicValues
-        );
+        (bytes8 n_inv_sqrt, bytes8 n1_inv, bytes8 s2, bytes8 n_bytes, bytes32 digest) =
+            rvVerifier.verifyRvProof(fixture.proof, fixture.publicValues);
         // Check that s is within error bounds of the proven s2 value
-        uint s = uint(fixture.s.abs());
+        uint256 s = uint256(fixture.s.abs());
         uint256 error = 2 * s + 1;
         uint256 s2_test = s * s >> fraction_bits;
         uint256 s2_256 = uint256(uint64(s2));
 
         assert(s2_test - error <= s2_256 && s2_256 <= s2_test + error);
-       
-        uint n_inv_sqrt_256 = uint256(uint64(n_inv_sqrt));
-        uint n1_inv_256 = uint256(uint64(n1_inv));
+
+        uint256 n_inv_sqrt_256 = uint256(uint64(n_inv_sqrt));
+        uint256 n1_inv_256 = uint256(uint64(n1_inv));
 
         // Check that n1_inv * (n - 1) is within error bounds of 1
-        uint n1 = uint(uint64(n_bytes))-1;
+        uint256 n1 = uint256(uint64(n_bytes)) - 1;
         console.log("n1: ", n1);
-        error =  2 * n_inv_sqrt_256 + 1;
+        error = 2 * n_inv_sqrt_256 + 1;
         console.log("error: ", error);
-        uint n1_inv_test = uint(uint64(n1_inv)) * n1 >> fraction_bits;
+        uint256 n1_inv_test = uint256(uint64(n1_inv)) * n1 >> fraction_bits;
         console.log("n1_inv_test: ", n1_inv_test);
-        uint one_256 = 1 << fraction_bits;
+        uint256 one_256 = 1 << fraction_bits;
         console.log("one_256: ", one_256);
         assert(one_256 - error <= n1_inv_test && n1_inv_test <= one_256 + error);
-        
+
         // Check that n_inv_sqrt * n_inv_sqrt * n is within error bounds of 1
-        uint n = uint(uint64(n_bytes));
+        uint256 n = uint256(uint64(n_bytes));
         console.log("n: ", n);
-        uint n_inv_sqrt_test = n_inv_sqrt_256 * n_inv_sqrt_256 * n >> 2 * fraction_bits;
+        uint256 n_inv_sqrt_test = n_inv_sqrt_256 * n_inv_sqrt_256 * n >> 2 * fraction_bits;
         console.log("n_inv_sqrt_test: ", n_inv_sqrt_test);
         assert(one_256 - error <= n_inv_sqrt_test && n_inv_sqrt_test <= one_256 + error);
     }
-    
+
     function testFail_InvalidRvProof() public view {
         SP1ProofFixtureJson memory fixture = loadFixture();
